+++ packages/react-components/src/index.ts
@@ -3,5 +3,6 @@ export * from './useConnection';
 export * from './useContractSelector';
 export * from './useDisconnect';
 export * from './useConnectorTypeStatus';
+export * from './useGrpcClient';
 export * from './WithWalletConnector';
 export * from '@concordium/wallet-connectors';