+++ samples/contractupdate/src/Root.tsx
@@ -1,18 +1,18 @@
 import React, { useEffect, useState } from 'react';
 import { Alert, Button, Col, Container, Row, Spinner } from 'react-bootstrap';
 import {
+    MAINNET,
+    TESTNET,
     WalletConnectionsProps,
     WithWalletConnector,
-    useConnect,
-    useConnection,
     useDisconnect,
-    withJsonRpcClient,
 } from '@concordium/react-components';
+import { useConnect, useConnection, useGrpcClient } from '@concordium/react-components';
 import { App } from './App';
 import { ConnectedAccount } from './ConnectedAccount';
 import { NetworkSelector } from './NetworkSelector';
 import { WalletConnectorButton } from './WalletConnectorButton';
-import { BROWSER_WALLET, MAINNET, TESTNET, WALLET_CONNECT } from './config';
+import { BROWSER_WALLET, WALLET_CONNECT } from './config';
 import { errorString } from './util';
 
 export default function Root() {
@@ -41,13 +41,12 @@ function Main(props: WalletConnectionsProps) {
 
     const [rpcGenesisHash, setRpcGenesisHash] = useState<string>();
     const [rpcError, setRpcError] = useState('');
+    const rpc = useGrpcClient(network);
     useEffect(() => {
-        if (connection) {
+        if (rpc) {
             setRpcGenesisHash(undefined);
-            withJsonRpcClient(connection, async (rpc) => {
-                const status = await rpc.getConsensusStatus();
-                return status.genesisBlock;
-            })
+            rpc.getConsensusStatus()
+                .then((status) => status.genesisBlock)
                 .then((hash) => {
                     setRpcGenesisHash(hash);
                     setRpcError('');
@@ -57,7 +56,7 @@ function Main(props: WalletConnectionsProps) {
                     setRpcError(errorString(err));
                 });
         }
-    }, [connection, genesisHash, network]);
+    }, [rpc]);
     return (
         <>
             <Row className="mt-3 mb-3">
@@ -98,7 +97,7 @@ function Main(props: WalletConnectionsProps) {
             </Row>
             <Row className="mt-3 mb-3">
                 <Col>
-                    <ConnectedAccount connection={connection} account={account} network={network} />
+                    <ConnectedAccount network={network} rpc={rpc} account={account} />
                 </Col>
             </Row>
             <Row className="mt-3 mb-3">
@@ -111,7 +110,7 @@ function Main(props: WalletConnectionsProps) {
                         />
                     )}
                     {rpcError && <Alert variant="warning">RPC error: {rpcError}</Alert>}
-                    <App network={network} connection={connection} connectedAccount={account} />
+                    <App network={network} rpc={rpc} connection={connection} connectedAccount={account} />
                 </Col>
             </Row>
         </>