+++ samples/sign-message/src/App.tsx
@@ -2,6 +2,7 @@ import { ResultAsync, err, ok } from 'neverthrow';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { Alert, Button, Col, Container, Form, InputGroup, Row, Spinner } from 'react-bootstrap';
 import {
+    TESTNET,
     WalletConnectionsProps,
     WithWalletConnector,
     binaryMessageFromHex,
@@ -13,7 +14,7 @@ import {
 } from '@concordium/react-components';
 import { AccountTransactionSignature } from '@concordium/web-sdk';
 import { WalletConnectorButton } from './WalletConnectorButton';
-import { BROWSER_WALLET, TESTNET, WALLET_CONNECT } from './config';
+import { BROWSER_WALLET, WALLET_CONNECT } from './config';
 import { errorString } from './util';
 
 export default function App() {